docs: updated README in animations_sprites folder; renamed manim_sprites.py file
