+++ src/animations_sprites/manim_sprites.py
